requirements.txt
streamlit>=1.28.0
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
